[opt](test) Merge 12 tiny nereids test classes into 3 suites
Every FE unit test class pays a full FE startup. Measured on the FE UT CI
agent (build 1008420), that fixed cost is about 54s per class: ~16s of JVM
fork and class loading (measured from test classes that never start an FE)
plus ~38s of FE startup.

210 of the 214 nereids test classes that start an FE finish in under 90s,
i.e. they are almost entirely that fixed cost. Merging classes that already
share the exact same fixture removes it without changing any test.

This first batch merges 12 classes that all extend SqlTestBase and override
no setup at all, so they already ran against an identical fixture (database
"test" with tables T0..T4):

  sqltest/SqlPlanSuiteTest                  <- CascadesJoinReorderTest, InferTest,
                                               JoinTest, MultiJoinTest, SortTest
  rules/exploration/mv/MvExplorationSuiteTest <- BuildStructInfoTest, EliminateJoinTest,
                                               HyperGraphAggTest, HyperGraphComparatorTest,
                                               NullRejectInferenceTest
  rules/rewrite/RewriteRuleSuiteTest        <- PruneOlapScanTabletTest,
                                               EliminateConstHashJoinConditionTest

Test bodies are moved verbatim; method names are unchanged so test reports
keep the same identifiers. Each merged file keeps a "from <OriginalClass>"
section header, so grepping the old class name still lands on its tests.

Verified: 53 tests before, 53 tests after, all passing.

Deliberately not merged here:
- nereids/mv (6 classes): they create materialized views with colliding names
  inside their test methods, so sharing one FE makes them interfere. Needs
  namespace isolation first.
- jobs/joinorder/hypergraph (4 classes): testPullUpQueryFilter exists with
  different bodies in two of them.
- Classes with real work (DistributeHintTest 764s, OtherJoinTest 256s,
  GraphSimplifierConsistencyTest 223s, PreMaterializedViewRewriterTest 205s)
  are intentionally left alone: merging them would create a straggler that
  hurts wall clock under the 12-way parallel CI run.
